backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for the project"
author = "Shashank Rajak"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[google-genai]>=0.3.26",
    "langchain-community>=0.3.26",
    "python-dotenv>=1.1.1",
    "langgraph>=0.5.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "langchain-chroma>=0.2.4",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
